requirements.txt
openai-agents
python-dotenv
mysql-connector-python